test(app): split page scenarios
